+++ src/slices/authSlice.ts
@@ -22,6 +22,6 @@ const authSlice = createSlice({
 export const { setAccessToken } = authSlice.actions;
 // export const selectAccessToken = (state: rootState) => state.auth.accessToken;
 export const selectAccessToken =
-  "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ1bmhoeXllZTIyQGdtYWlsLmNvbSIsImlhdCI6MTY5MzM4NDQ4MywiZXhwIjoxNjkzNDcwODgzfQ.F7ud7-e9OsAj7uAc9h8DakGLpZkxIg_jm6OqvvcxShI";
+  "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJ1bmhoeXllZTIyQGdtYWlsLmNvbSIsImlhdCI6MTY5MzQ3MTI1MywiZXhwIjoxNjkzNTU3NjUzfQ.5WryYuj_lvMQoPwshkI0czXEUcNMH7ubSUNcfLjPHJ4";
 
 export default authSlice.reducer;
